separate rx and tx start and stop functions. start adding phy read function.
