+++ examples/blink.rs
@@ -14,12 +14,7 @@ use embedded_hal::serial::Read as ReadHal;
 use tm4c129x_hal::gpio::GpioExt;
 use tm4c129x_hal::serial;
 use tm4c129x_hal::time::Bps;
-
-use catnip::{IPV4Addr, MACAddr};
-use tm4c129_launchpad::{
-    board,
-    drivers::ethernet::{socket::UDPSocket, RXBUFSIZE},
-};
+use tm4c129_launchpad::board;
 
 #[no_mangle]
 pub fn stellaris_main(mut board: board::Board) -> ! {
@@ -37,21 +32,6 @@ pub fn stellaris_main(mut board: board::Board) -> ! {
     );
     let mut delay = tm4c129x_hal::delay::Delay::new(board.core_peripherals.SYST, board::clocks());
 
-    let udp = UDPSocket {
-        src_macaddr: MACAddr {
-            value: board.enet.src_macaddr,
-        },
-        src_ipaddr: IPV4Addr {
-            value: [10, 0, 0, 2],
-        },
-        src_port: 8053,
-        dst_macaddr: None,
-        dst_ipaddr: IPV4Addr {
-            value: [172, 17, 0, 1],
-        },
-        dst_port: 8053,
-    };
-
     uart.write_all("Welcome to Launchpad Blink\n");
     let mut loops = 0;
     loop {
@@ -60,63 +40,7 @@ pub fn stellaris_main(mut board: board::Board) -> ! {
         while let Ok(ch) = uart.read() {
             // Echo
             writeln!(uart, "byte read {}", ch).unwrap_or_default();
-
-            // Show MAC address
-            let addr = udp.src_macaddr.value;
-            writeln!(
-                uart,
-                "MAC Address: {:x}:{:x}:{:x}:{:x}:{:x}:{:x}",
-                addr[0], addr[1], addr[2], addr[3], addr[4], addr[5]
-            )
-            .unwrap_or_default();
-
-            // Debugging
-
-            // Test ethernet receive (without UDP socket)
-            let mut buf = [0_u8; RXBUFSIZE];
-            unsafe {
-                match &board.enet.receive(&mut buf) {
-                    Ok(num_bytes) => {
-                        writeln!(uart, "Received {} ethernet bytes", num_bytes).unwrap_or_default()
-                    }
-                    Err(x) => writeln!(uart, "Ethernet RX error: {:?}", x).unwrap_or_default(),
-                };
-            }
-            let rxdl = &mut (board.enet.rxdl);
-            writeln!(uart, "{:?}", rxdl).unwrap_or_default();
         }
-
-        // Test UDP transmit
-        match udp.transmit::<3>(&mut board.enet, *b"hello world!") {
-            Ok(_) => writeln!(uart, "UDP transmit started").unwrap_or_default(),
-            Err(x) => writeln!(uart, "UDP TX error: {:?}", x).unwrap_or_default(),
-        };
-        let txdl = &mut (board.enet.txdl);
-        writeln!(uart, "{:?}", txdl).unwrap_or_default();
-
-        // Check EMAC status
-        let status = board.enet.emac.status.read().txpaused().bit_is_set();
-        writeln!(uart, "EMAC TX paused? {status}");
-        let status = board.enet.emac.hostxdesc.read().bits();
-        writeln!(uart, "EMAC TX descr addr {status}");
-        let status = board.enet.emac.hostxba.read().bits();
-        writeln!(uart, "EMAC TX buf addr {status}");
-        let status = board.enet.emac.txdladdr.read().bits();
-        writeln!(uart, "EMAC TXDL addr {status}");
-        let status = board.enet.emac.status.read().txfe().bit_is_set();
-        writeln!(uart, "EMAC TX FIFO not empty? {status}");
-        let status = board.enet.emac.status.read().twc().bit_is_set();
-        writeln!(uart, "EMAC TX FIFO write controller active? {status}");
-        let status = board.enet.emac.status.read().tpe().bit_is_set();
-        writeln!(uart, "EMAC MII transmit protocol engine status? {status}");
-        let status = board.enet.emac.status.read().tfc().variant();
-        writeln!(uart, "EMAC MII transmit frame controller status {status:?}");
-        let status = board.enet.emac.dmaris.read().ts().bits();
-        writeln!(uart, "EMAC DMA transmit process state {status:?}");
-        let status = board.enet.emac.dmaris.read().fbi().bit_is_set();
-        writeln!(uart, "EMAC DMA bus error? {status:?}");
-        let status = board.enet.emac.dmaris.read().ae().bits();
-        writeln!(uart, "EMAC DMA access error type {status:?}");
 
         loops = loops + 1;
