+++ examples/flash.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+INFILE="target/thumbv7em-none-eabihf/release/examples/$1" ;
+OUTFILE="target/thumbv7em-none-eabihf/release/examples/$1.bin" ; 
+
+# Compile
+cargo build --example $1 --release
+# Convert binary
+arm-none-eabi-objcopy -O binary $INFILE $OUTFILE
+# Flash
+sudo $LM4FLASH $OUTFILE
+# Show binary size
+ls -lh $OUTFILE
